requirements.txt
appdirs==1.4.0
docutils==0.13.1
lockfile==0.12.2
packaging==16.8
paho-mqtt==1.2
pycups==1.9.73
pyparsing==2.1.10
python-daemon==2.1.2
six==1.10.0
